+++ CHANGELOG.md
@@ -51,7 +51,9 @@ All notable changes to the "nwscript-ee-language-server" extension will be docum
 
 ## [1.5.3]
 
-- Goto will now work for functions from `nwscript.nss` if the file is in your project.
+- Goto will now work for functions and constants from `nwscript.nss` if the file is in your project.
+- Fixed a few issues with the tokenizer.
+- Fixed a small issue with `const` expressions resolution.
 - Fixed the compilation provider not reporting warnings.
 - New compiler setting `reportWarnings`. True by default.
 - New formatter setting `verbose`. False by default.